refactor: `client` file structure
