+++ src/client/utils/get-query-string.ts
@@ -1,4 +1,4 @@
-import { IButton } from '../../types/interfaces';
+import { IButton } from '../types/interfaces';
 
 export const getQueryString = (query: IButton[]): string => {
   return query.map((elem) => elem.char).join('');
