+++ tsconfig.json
@@ -10,5 +10,5 @@
     "jsx": "react-jsx"
   },
   "include": ["src/**/*"],
-  "exclude": ["node_modules", "src/client/**/*"]
+  "exclude": ["node_modules"]
 }
